Fixed pr to not be chaos
Added persistent SQLite table initialization, implemented registration and login with simple MD5 hashing, and added deposit/withdrawal transactions using double values. Made entity fields mutable with setters, removed the automatic database reset and custom exception hierarchy, and added the local database file to .gitignore.

+++ src/main/java/dragon/repository/WithdrawalTransactionRepository.java
@@ -13,7 +13,7 @@ public class WithdrawalTransactionRepository {
         try (PreparedStatement statement = connection.prepareStatement(sql)) {
             statement.setString(1, transaction.getId().toString());
             statement.setString(2, transaction.getUserId().toString());
-            statement.setBigDecimal(3, transaction.getAmount());
+            statement.setDouble(3, transaction.getAmount());
             statement.setString(4, transaction.getDate().toString());
             statement.executeUpdate();
         }
